requirements.txt
fastapi
uvicorn
numpy
Pillow
imgaug
jinja2
python-multipart